Move splunkFeeder to sender package
Drop the temporary splunk package
